Move basic functionality into OSCUtils
